Run load test with auth on
